IMR-173 refactor: change import path
